add makeEnv; fix NTP tests
