+++ tests/unit_tests/test_ntp.js
@@ -9,6 +9,7 @@ const debug = require("debug")(config.debugPrefix + "unittests:ntp");
 
 const ntpServer = process.env.BEAME_TESTS_NTP_SERVER || "pool.ntp.org";
 const ntpServerPort = process.env.BEAME_TESTS_NTP_SERVER_PORT || 123;
+const range = Number.parseInt(process.env.BEAME_TEST_NTP_RANGE || '10');
 
 describe('ntp', () => {
 	it('check date', async () => {
@@ -20,8 +21,8 @@ describe('ntp', () => {
 		debug("Current machine time is: ", local);
 		assert(local);
 
-		const diff =  (date.getTime() - local.getTime()) / 1000;
+		const diff =  Math.abs((date.getTime() - local.getTime()) / 1000);
 		debug("diff is : ",diff);
-		assert(-0.1 > diff < 0.1);
+		assert(diff < range);
 	});
 });
